pyproject.toml
[project]
name = "ai-companion"
version = "0.1.0"
description = "Your own AI buddy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=1.3.2",
    "elevenlabs>=1.50.3",
    "fastapi[standard]>=0.115.6",
    "groq>=0.13.1",
    "langchain-community>=0.3.13",
    "langchain-groq>=0.2.2",
    "langchain>=0.3.13",
    "pydantic==2.10.0",
    "together>=1.3.10",
    "langgraph>=0.2.60",
    "langchain-openai>=0.2.14",
    "pydantic-settings>=2.7.0",
    "pre-commit>=4.0.1",
    "supabase>=2.11.0",
    "langgraph-checkpoint-duckdb>=2.0.1",
    "duckdb>=1.1.3",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "qdrant-client>=1.12.1",
    "sentence-transformers>=3.3.1",
    "httpx>=0.27.0",  
    "pillow>=10.0.0"
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E501"]
